feat: rename the name of services for storage to provider
