+++ libs/storage/src/providers/index.ts
@@ -1,3 +1,3 @@
-export * from './s3-storage.service';
-export * from './rustfs-storage.service';
-export * from './local-fs-storage.service';
+export * from './s3-storage.provider';
+export * from './rustfs-storage.provider';
+export * from './local-fs-storage.provider';
