+++ libs/storage/src/storage.module.ts
@@ -1,8 +1,8 @@
+import { LocalFsStorageService } from './providers/local-fs-storage.provider';
 import { Module } from '@nestjs/common';
+import { RustFsStorageService } from './providers/rustfs-storage.provider';
+import { S3StorageService } from './providers/s3-storage.provider';
 import { StorageService } from './storage.service';
-import { S3StorageService } from './providers/s3-storage.service';
-import { RustFsStorageService } from './providers/rustfs-storage.service';
-import { LocalFsStorageService } from './providers/local-fs-storage.service';
 
 @Module({
   providers: [StorageService, S3StorageService, RustFsStorageService, LocalFsStorageService],
